// frontend/src/types/purchase.types.ts
// Defines the TypeScript structure for a purchase, including the seller, product, quantity, pricing, payment, and creation date.

export interface IPurchase {
  _id: string
  user: string
  seller: string
  product: string
  sellerName: string
  productName: string
  quantity: number
  unitPrice: number
  totalPrice: number
  paid: number
  createdAt: string
}